+++ apps/ledger-live-mobile/src/dynamicContent/dynamicContent.tsx
